requirements.txt
aiohttp==2.0.7
appdirs==1.4.3
async-timeout==1.2.1
chardet==3.0.3
gunicorn==19.7.1
mock==2.0.0
multidict==2.1.5
packaging==16.8
pbr==3.0.1
py==1.4.33
pyparsing==2.2.0
pytest==3.0.7
pytest-aiohttp==0.1.3
PyYAML==3.12
six==1.10.0
yarl==0.10.2
